[spaces.dg.fem] moved DiscontinuousLangrange -> DG, refs #14
